==> shc2021/1_cohort1_init.ipynb <==
This notebook was originally developed by Conor K. Corbin, modified by Minh Nguyen

### Description:
- Query the validation cohort with admitted ED patients
- Use shc_core_2021 to query 2020 and 2021 data
- similar to `1.1_cohort`

**Output**: 

validation cohort `1_cohort1`

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import os 
from datetime import datetime

from google.cloud import bigquery
from google.cloud.bigquery import dbapi

### THIS IS MEANT TO RUN ON NERO GCP Jupyter notebook
os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = '/home/jupyter/.config/gcloud/application_default_credentials.json'

# FOR LOCAL COMPUTER:
# os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = r'C:\Users\User\AppData\Roaming\gcloud\application_default_credentials.json' 

os.environ['GCLOUD_PROJECT'] = 'som-nero-phi-jonc101' 
%load_ext google.cloud.bigquery

from google.cloud import bigquery
client=bigquery.Client()

datadir6 = "../../DataTD/shc2021"
valdir = "../../OutputTD/shc2021"

pd.options.display.max_rows = 100

### Query ADT and return rows where patient encounter id is associated with an ED visit
* Get ER stays that BEGIN between 2015 and 2020. Some of our index times will be in 2021 because the ER visit will start on 12-31-2019 and ADT has 2021 data.
* For validation cohort, get between 2020 and 2021. Will filter to get only April 2020 to 2021 (or only take csn that are not already in previously used cohort).
* Assumes patient encounters associated with an ED visit START with an ED visit, which is probably always true-- But might want to check this. 
* Assumes we can trust pat_enc_csn_id_coded as a patient encounter - which we can't always.  Encounters are weird, and sometimes multiple pat_enc_csn_id_coded's will overlap in time for a particular patient (Stephen knows more about this). UPDATE - ER encounters should have unique csns. This was checked toward the end

query = """

WITH er_admit_times AS (
SELECT pat_enc_csn_id_coded, MIN(EXTRACT(YEAR FROM event_time_jittered_utc)) admit_year
FROM shc_core_2021.adt
WHERE pat_class_c = "112"
AND pat_service = "Emergency"
GROUP BY pat_enc_csn_id_coded
)

SELECT er.anon_id, er.pat_enc_csn_id_coded, er.effective_time_jittered_utc,
        er.base_pat_class_c, er.pat_lv_of_care, er.pat_class, er.event_type
           
FROM shc_core_2021.adt as er
INNER JOIN er_admit_times
USING (pat_enc_csn_id_coded)
WHERE er_admit_times.admit_year BETWEEN 2015 AND 2022
ORDER BY anon_id, pat_enc_csn_id_coded, effective_time_jittered_utc 
"""
query_job =client.query(query)
df=query_job.to_dataframe()

# order by: changed from event_time to effective_time

print(df.shape)
df.head() # 3630469

# saving the df - stored in box
df.to_csv("{}/dfer.csv".format(datadir), index=None)

# run this ONLY if read in the stored df locally
df = pd.read_csv("{}/dfer.csv".format(datadir))
print(df.shape)
df.head()

# change columns to datetime, if read from locally stored data (if queried directly from BQ, already in datetime format)
df.effective_time_jittered_utc = pd.to_datetime(df.effective_time_jittered_utc)

### Get counts on unique patient trajectories 
where patients were sent throughout their stay - this is at the granularity of inpatient vs observation vs discharged straight from ER - not the unit they are sent to.

Trajectories of patient encounter ids.
* Most get discharged directly from ER.
* Next most common is ER -> inpatient (which is what we're focusing on). 
* Next is ER to observation
* Next is ER to observation to inpatient etc. 

# Filter for rows where pat_class changes = meaning they don't get discharged directly from the ER
df_change = df[~df['base_pat_class_c'].isna()]
print(len(df)) # 3630469
print(len(df_change)) # 725732

# Collapse pat class on patient encounter id, create trajectories, group by trajectories, count patient encounters with those trajectories. 
pd.set_option("display.max_colwidth", None)

df_change[['pat_enc_csn_id_coded', 'pat_class']].groupby('pat_enc_csn_id_coded').agg(
{'pat_class' : lambda x: ' -> '.join([c for c in x])}).reset_index().groupby('pat_class').agg(
{'pat_enc_csn_id_coded' : 'count'}).reset_index().sort_values('pat_enc_csn_id_coded', ascending=False)

### Patient Level of Care Column is the indicator we want - 
It seems to be missing a lot 
Update: LOC not always on the same row as the change in pat code status ie inpatient vs observation vs emergency services

Filter for patients with an inpatient code immediately after emergency services and create trajectories

import datetime

def has_inpatient_code(arr):
    for a in arr:
        if a == 'Inpatient':
            return True
    return False

def has_inpatient_code_after_er(arr):
    """Assumes arr is ordered by time"""
    has_er = False
    for a in arr:
        if a == 'Emergency Services':
            has_er = True
        elif a == 'Inpatient' and has_er == True:
            return True
        else:
            has_er = False
        
    return False

def get_trajectory(arr):
    # Creates trajectory but only adds to path when level of care changes
    traj = []
    for i, a in enumerate(arr):
        if len(traj) == 0:
            traj.append(a)
        elif a != traj[-1]:
            traj.append(a)
    return ' -> '.join(traj)

# Get a set of patient_encounter_ids that have an inpatient code
df_temp = df_change[['pat_enc_csn_id_coded', 'pat_class']].groupby('pat_enc_csn_id_coded').agg(
{'pat_class' : has_inpatient_code_after_er}).reset_index()
inpatient_ids = set(df_temp[df_temp['pat_class'] == True]['pat_enc_csn_id_coded'].values)

# Filter original df for patients in this set and create level of care trajectories.
df_lofc = df[df['pat_enc_csn_id_coded'].isin(inpatient_ids)]

# Get df of csn_ids and admit timestamps
df_admit_times = df_change[df_change['pat_enc_csn_id_coded'].isin(inpatient_ids)]
df_admit_times = df_admit_times[df_admit_times['pat_class'] == 'Inpatient'].groupby(
    'pat_enc_csn_id_coded').first().reset_index()[['pat_enc_csn_id_coded', 'effective_time_jittered_utc']].rename(
    columns={'effective_time_jittered_utc' : 'admit_time_jittered'})
df_admit_times.head()

# Should now be querying adt again for anon_id that match each er CSN id and then look ahead 24 hours to mitigate overlapping csn issue

# Merge to df_lofc and create column called time_since_admit
df_lofc = pd.merge(df_lofc, df_admit_times, how='left', on='pat_enc_csn_id_coded')
df_lofc['time_since_admit'] = df_lofc.apply(lambda x: x.effective_time_jittered_utc - x.admit_time_jittered, axis=1)

# Filter df_lofc so that we only look 24 hours into admission
df_lofc = df_lofc[df_lofc['time_since_admit'] < datetime.timedelta(hours=24)]

print(len(df_lofc)) # 731347

### Print the trajectories 24 hours into admission

# Merge this to df_lofc and 
df_traj = df_lofc[['pat_enc_csn_id_coded', 'pat_lv_of_care']].dropna().groupby(
'pat_enc_csn_id_coded').agg({'pat_lv_of_care' : get_trajectory}).reset_index().groupby(
'pat_lv_of_care').count().reset_index().sort_values('pat_enc_csn_id_coded', ascending=False)

# Print cause these trajectories are long
for i in range(len(df_traj)):
    print(df_traj['pat_lv_of_care'].values[i], ' : ', df_traj['pat_enc_csn_id_coded'].values[i])

# Count number of encounters with a trajectory. 
print('Number of encounters with a trajectory : ', df_traj['pat_enc_csn_id_coded'].sum())
print('Total Number of inpatient encounters : ', len(inpatient_ids))

### Create Labelling function for highest level of care with 24 hours of admit
For each csn id, we'll create positive or negative labels based on whether within 24 hours of admit they have a critical care label in pat_lv_of_care.  This means that if they are originally place in critical care but then sent to acute care we'll still label them as crit care.  Label is thus the max level of care within 24 ours of admit

Output dataframe should have anon_id, csn_id, admit_time, label

NOTE : this isn't completley correct because i've grouped on pat_enc_csn_id_coded.  I really should be taking the anon_id from each er csn id and looking ahead 24 hours in the adt table to see if there exist other csn id's associated with the encounter where the level of care changes... 

UPDATE: OK to group by pat_enc_csn_id_coded, take the first admit time for each visit, first anon_id only matter if there are multiple anon_ids with a CSN. This is not the case in our data, checked.

def was_placed_in_critical_care(arr):
    """Returns true if patient placed in crtical care within 24 hours of admit
       Assumes we have already done the 24 hours logic
       Assumes no overlapping csn ids... """
    for a in arr:
        if a == 'Critical Care':
            return 1
    return 0

# group by anon_id and csn
df_labels = df_lofc.groupby(['anon_id','pat_enc_csn_id_coded']).agg({ #cohort0_all_anon_ids
    'admit_time_jittered' : 'first',
    'pat_lv_of_care' : was_placed_in_critical_care}).rename(
    columns={"pat_lv_of_care" : 'label'}).reset_index()[['anon_id', 'pat_enc_csn_id_coded', 'admit_time_jittered', 'label']]

print(len(df_labels)) # 55168

# Save to .csv and read back as necessary
df_labels.to_csv(os.path.join(valdir,"1_cohort1.csv"), index=False)

# check labels
df_labels.groupby('label').count() # 85.7%

# make sure you only see 1's in the anon_id column!! (or else there are multiple anon_ids for a csn)
df_labels[['anon_id','pat_enc_csn_id_coded']].groupby('pat_enc_csn_id_coded').nunique().reset_index().sort_values('anon_id')

# don't want to use this one anymore
# group by csn and take first anon_id only, to avoid 1 csn with multiple anon_id
df_labels2 = df_lofc.groupby('pat_enc_csn_id_coded').agg({ #cohort0
    'admit_time_jittered' : 'first',
    'pat_lv_of_care' : was_placed_in_critical_care}).rename(
    columns={"pat_lv_of_care" : 'label'}).reset_index()[['anon_id', 'pat_enc_csn_id_coded', 'admit_time_jittered', 'label']]

print(len(df_labels2)) # 55383 vs. 55402
df_labels2.groupby('label').count() #55723 for df vs. 55383 for df_change

# Save to .csv and read back as necessary
df_labels2.to_csv(os.path.join(datadir, 'cohort0.csv'), index=None)

### Save to big query, the orginial 1_1_cohort = cohort _init_

# read back as necessary
df_labels = pd.read_csv(os.path.join(valdir, '1_cohort1.csv'), index_col=False)
print(len(df_labels))
print(list(df_labels.columns)) #75455

table_schema = [{'name' : 'anon_id', 'type' : 'STRING'},
                {'name' : 'pat_enc_csn_id_coded', 'type' : 'INTEGER'},
                {'name' : 'admit_time_jittered', 'type' : 'TIMESTAMP'},
                {'name' : 'label', 'type' : 'INTEGER'}]
                       
DATASET_NAME = 'triageTD'
TABLE_NAME = '1_cohort1'
df_labels.to_gbq(destination_table='triageTD.%s' % TABLE_NAME,
                 project_id='som-nero-phi-jonc101',
                 table_schema=table_schema,
                 if_exists='replace')
