==> shc2021/6.7_cohort4_R.ipynb <==
### Description: NEW
Refine new cohort4 to only include CSNs that are not already in the original cohort
- New cohort: shc_core_2021, from 04/2020 - 2021
- Original old cohort: shc_core, 2015 - 03/2020

However, some CSNs in the new cohort are also in the old cohort, but with different anon_id and admit time
- Remove these overlapping CSNs. This is clean if use both cohort4 from old and new data
- When using Demographics/HW data later in 6.8 notebook, be aware that there are a few overlapped CSNs. This is due to cohort2 was used to get demo/HW data. These overlapped CSNs were actually removed from cohort3 already. They are removed from new cohort4 again when merging with the old demo/HW data to prevent further issues. But this `cohort4` remain intact.

**Input:**
- `1_4_cohort.csv` (original cohort4)
- `6_7_0_cohort4` (new cohort4 with labels)


**Output:**
- `6_7_cohort4` size 60,464. This is the *final cohort* combing 2015 - 03/2020 (43,980) and 04/2020 - 2021 (16,484) data


library(data.table)
library(tidyverse)
library(lubridate)
# library(Matrix)
# library(slam)
# library(bit64)
# library(mtools) for one hot coder, not available on Nero or use caret or tidyr
options(repr.matrix.max.rows=200, repr.matrix.max.cols=30)

datadir = "../../DataTD"
datadir6 = "../../DataTD/validation"
valdir = "../../OutputTD/6_validation"
cohortdir = "../../OutputTD/1_cohort"
featuredir = "../../OutputTD/2_features"

# old cohort up to 03/2020
cohort0 <- read.csv(file.path(cohortdir,  '1_4_cohort.csv'))
nrow(cohort0) #43980

# new cohort from 04/2020 - 2021
cohort6 <- read.csv(file.path(valdir,  '6_7_0_cohort4.csv'))
nrow(cohort6) #16700

# check overlapping cohort
length(setdiff(cohort0$pat_enc_csn_id_coded, cohort6$pat_enc_csn_id_coded))
length(setdiff(cohort6$pat_enc_csn_id_coded, cohort0$pat_enc_csn_id_coded)) 

# csn in cohort4 but not in 0
csn_new <- setdiff(cohort6$pat_enc_csn_id_coded, cohort0$pat_enc_csn_id_coded)
head(csn_new)

# merge with unique csn
cohort6 <- cohort6 %>% filter(pat_enc_csn_id_coded %in% csn_new)
nrow(cohort6)
length(unique(cohort6$pat_enc_csn_id_coded))

cohort <- bind_rows(cohort0, cohort6)
nrow(cohort)
nrow(cohort %>% select(anon_id) %>% distinct()) # 60464
nrow(cohort %>% select(pat_enc_csn_id_coded) %>% distinct()) # 60464
nrow(cohort %>% select(anon_id, pat_enc_csn_id_coded) %>% distinct()) # 60464
head(cohort)

write.csv(cohort6, file = file.path(valdir, "6_7_cohort4.csv"), row.names=FALSE) 
write.csv(cohort, file = file.path(valdir, "6_7_cohort4_all.csv"), row.names=FALSE)
