==> shc2021/10_notes.ipynb <==
# Notes

Check patient notes and create features:

## Dispo to 24hr

- count how many RN vs MD notes (author)
- count total all notes from anyone

import pandas as pd
import numpy as np
from datetime import datetime
from datetime import timedelta
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats

# view data frame

# adjust this to hide ID columns for posting to github
forrepo=True

def view_df(df):
    hidecols = []
    if forrepo:
        hidecols=['anon_id', 'pat_enc_csn_id_coded', 'inpatient_data_id_coded'] # these are hidden columns
        
    return(df.drop(hidecols, axis=1, errors='ignore').head())

# Data

Load in the data

Grabbed the adt table with the following code:

select *
    from `som-nero-phi-jonc101.shc_core_2021.adt` as adt
    where adt.pat_enc_csn_id_coded 
    in (select distinct pat_enc_csn_id_coded
        from `som-nero-phi-jonc101.triageTD.7_cohort4_3hr_labels_noOR`
    )

# read in data files
datadir = "../../DataTD/shc2021/"
savedir = "../../OutputTD/shc2021/"


adt_file = datadir + "cohort3_adt_2021.csv"
cohort_file = savedir + "7_cohort4_3hr_labels_noOR.csv"


full_adt = pd.read_csv(adt_file)
cohort = pd.read_csv(cohort_file)

print(full_adt.pat_enc_csn_id_coded.nunique())
print(cohort.pat_enc_csn_id_coded.nunique())

full_adt.dtypes

cohort.dtypes

# Bring in the ADT Table

We can use the ADT table to look at 

- pat_class
- pat_lv_of_care

# filter adt table down to test cohort
adt = full_adt[full_adt.pat_enc_csn_id_coded.isin(cohort.pat_enc_csn_id_coded)]

# these should match - it does!
print(adt.pat_enc_csn_id_coded.nunique())
print(cohort.pat_enc_csn_id_coded.nunique())

view_df(adt)

# add the admit time column from the prediction results csv to the adt csv
adt_admit = adt.merge(cohort[['pat_enc_csn_id_coded', 'admit_time']], how='left', on=['pat_enc_csn_id_coded'])

adt_admit.columns
view_df(adt_admit[['pat_enc_csn_id_coded', 'effective_time_jittered_utc', 
                 'pat_class', 'admit_time']])

adt_admit.columns

# get time of first emergency services event
first_ED = adt_admit[(adt_admit.pat_class == 'Emergency Services')].sort_values(['pat_enc_csn_id_coded', 'seq_num_in_enc']).groupby('pat_enc_csn_id_coded').first().reset_index()
print(first_ED.seq_num_in_enc.unique())

view_df(first_ED)

# first seq number include those Observation
first_ED = adt_admit[(adt_admit.seq_num_in_enc == 1)].sort_values(['pat_enc_csn_id_coded', 'seq_num_in_enc']).groupby('pat_enc_csn_id_coded').first().reset_index()
print(first_ED.pat_class.unique())

view_df(first_ED)

print(adt_admit.pat_enc_csn_id_coded.nunique())
print(first_ED.pat_enc_csn_id_coded.nunique())

set(adt_admit.pat_enc_csn_id_coded) - set(first_ED.pat_enc_csn_id_coded)

print("\n", first_ED.pat_service.value_counts())
print("\n", first_ED.event_type.value_counts())

first_ED['first_ed_time'] = first_ED.effective_time_jittered_utc

first_ED_time = first_ED[['pat_enc_csn_id_coded', 'first_ed_time']].drop_duplicates()
first_ED_time.head()

adt_ed = adt_admit.merge(first_ED_time, how='left')
view_df(adt_ed)

# save some of these columns to upload to BQ so we can query the notes table
sub_first_ED = first_ED[['anon_id', 'admit_time', 'first_ed_time']].drop_duplicates()
# sub_first_ED.to_csv("{}/10_cohort4_3hr_noOR_first_ed_time.csv".format(savedir), index=False)

# Pull the notes from BQ

I used the "7_cohort4_3hr_noOR_first_ED_time.csv" file to query the notes table on BQ. We need to do this since CSN does not match between the cohort and the notes table. I used the following code to get all notes that occur for the anon_id between door to 24hr after admit.

select *
    from `som-nero-phi-jonc101.shc_core_2021.clinical_doc_meta` as notes
    inner join 
        (   select anon_id, admit_time, first_ED_time
            from `som-nero-phi-jonc101.triageTD.7_cohort4_3hr_noOR_first_ED_time`) as cohort
    using (anon_id)
    where timestamp_add(cohort.admit_time, interval 24 HOUR) >= notes.effective_time_jittered_utc
    and notes.effective_time_jittered_utc >= cohort.first_ED_time

# read the notes in
notes_file = "{}/cohort4_3hr_noOR_notes_2021.csv".format(datadir)

notes = pd.read_csv(notes_file)

# check anon id since csns no longer match
print(adt_admit.anon_id.nunique()) #35144
print(notes.anon_id.nunique()) #35144

notes.head(3)

## -- takes a while to run -- ##

# change the effective time to datetime since read in from csv
notes.note_date_jittered_utc = pd.to_datetime(notes.note_date_jittered_utc)

# change admit time
notes.admit_time = pd.to_datetime(notes.admit_time, utc=True)

# Remove unwanted notes

Remove notes according to the following:

- remove "Letter" and "Telephone Encounter", "Progress Note, Outpatient"
- remove Ambulatory == Y

# remove specific note types
print(notes.value_counts('note_type'))

# removing from these categories
remove_types = ['Letter', 'Telephone Encounter', 'Progress Note, Outpatient']

filtered_notes = notes[-(notes.note_type.isin(remove_types))]

print("\n", filtered_notes.value_counts('note_type'))

# remove ambulatory == Y
print("\n", filtered_notes.value_counts('ambulatory')) 

filtered_notes_noAmbulatory = filtered_notes[filtered_notes.ambulatory != 'Y']

print("\n", filtered_notes_noAmbulatory.value_counts('ambulatory'))

# adjust the prover id columns

# assign auth_lnked_prov_map_id, 
# use auth_prov_map_id only if auth_lnked_prov_map_id is null
filtered_notes_noAmbulatory['prov_map_id'] = np.where(filtered_notes_noAmbulatory['auth_lnked_prov_map_id'].isnull(),
                                                     filtered_notes_noAmbulatory['author_prov_map_id'],
                                                     filtered_notes_noAmbulatory['auth_lnked_prov_map_id'])


print(filtered_notes_noAmbulatory.auth_lnked_prov_map_id.isna().sum())
print(filtered_notes_noAmbulatory.author_prov_map_id.isna().sum())
print(filtered_notes_noAmbulatory.prov_map_id.isna().sum()) # 841 notes do not have an author id

print(filtered_notes_noAmbulatory.prov_map_id.nunique())

# Treatment team

Bring in the treatment team data so we know classify notes as written by MD or RN.

I used the following SQL to query the treatment team table in BQ to get those that overlap with out notes.


select distinct name, prov_map_id
    from `som-nero-phi-jonc101.shc_core_2021.treatment_team` as team
    where team.prov_map_id
    in 
    (select distinct notes.auth_lnked_prov_map_id
        from `som-nero-phi-jonc101.shc_core_2021.clinical_doc_meta` as notes
        inner join 
            (   select anon_id, admit_time, first_ED_time
                from `som-nero-phi-jonc101.triageTD.7_cohort4_3hr_noOR_first_ED_time`) as cohort
        using (anon_id)
        where timestamp_add(cohort.admit_time, interval 24 HOUR) >= notes.effective_time_jittered_utc
        and notes.effective_time_jittered_utc >= cohort.first_ED_time
    ) 
union distinct

select distinct name, prov_map_id
    from `som-nero-phi-jonc101.shc_core_2021.treatment_team` as team
    where team.prov_map_id
    in 
    (select distinct notes.auth_lnked_prov_map_id
        from `som-nero-phi-jonc101.shc_core_2021.clinical_doc_meta` as notes
        inner join 
            (   select anon_id, admit_time, first_ED_time
                from `som-nero-phi-jonc101.triageTD.7_cohort4_3hr_noOR_first_ED_time`) as cohort
        using (anon_id)
        where timestamp_add(cohort.admit_time, interval 24 HOUR) >= notes.effective_time_jittered_utc
        and notes.effective_time_jittered_utc >= cohort.first_ED_time
    ) 

# read in the treatment team file
treatment_team_file = "{}/cohort4_3hr_noOR_team_2021.csv".format(datadir)
team = pd.read_csv(treatment_team_file)# [['name', 'prov_map_id']]
print(team.shape)

team.head()

# connect the treatment team to the notes table
notes_team = filtered_notes_noAmbulatory.merge(team, how='left')
print(notes_team.name.value_counts())

notes_team.head(3)

# add columns for the author name
rn = ['Registered Nurse', 'Resource Nurse', 'Float Nurse', 
      'Triage Nurse', 'Licensed Vocational Nurse', 'Post-Transplant Nurse', 
      'Nursery Nurse', 'Delivery Nurse', 'Specialty Nurse']
md = ['Primary Team', 'Emergency Resident', 'Primary Resident', 'Primary Intern', 'Senior Resident', 
        'Primary Advanced Practice Provider', 'Nurse Practitioner', 'Intern', 'Resident', 'Physician Assistant', 
         'Cardiologist', 'Primary Fellow', 'Pulmonologist', 'Fellow', 'Primary Sub-intern', 'Sub-Intern', 
         'Gastroenterologist', 'Infectious Disease', 'Chief Resident', 'Co-Attending', 'Primary Care Physician',  
         'Hematologist', 'Internist', 'Surgeon', 'Medical Oncologist', 'Dermatologist', 'Surgical Oncologist', 
         'Endocrinologist', 'Nephrologist', 'Trauma Attending', 'Diabetes MD', 'Trauma Resident', 
         'BMT Attending Provider', 'Urologist', 'Transplant Pulmonologist', 'Neurologist', 'Anesthesiologist', 
         'Primary Physician Assistant', 'Hepatologist', 'Transplant Surgeon', 'Obstetrician', 
         'Family Practitioner', 'Heart Failure Cardiologist', 'Radiation Oncologist', 'Referring Pulmonologist', 'Transplant Nephrologist',
         'Cross Cover Intern', 'Cross Cover Sub-Intern', 'Cross Cover Attending', 'Cross Cover Resident', 
         'Cross Cover Fellow', 'Cross Cover Advanced Practice Provider',
         'Consulting Service', 'Consulting Attending', 'Consulting Fellow', 'Consulting Resident', 'Consulting Intern', 
         'Psychologist', 'Consulting Medical Oncologist', 'Consulting Hematologist', 'Consulting Surgical Oncologist', 'Post-Transplant Nephrologist']


# assign the author type
notes_team['author_type'] = np.where(notes_team['name'].isin(rn), 'rn', None)
notes_team['author_type'] = np.where(notes_team['name'].isin(md), 'md', notes_team['author_type'])


print(notes_team.author_type.value_counts()) # md 789781, rn 540917

first_ED.columns

# Connect cohort CSN to notes

Connect the cohort CSN to the notes.

# get the csn and important columns from first ED table
cols = ['pat_enc_csn_id_coded', 'anon_id', 'admit_time', 'first_ed_time']
csns = first_ED[cols]

# chage admit time to datetime
csns.admit_time = pd.to_datetime(csns.admit_time, utc=True)
csns.head()

# join the notes table by anon_id
csn_notes = csns.merge(notes_team, how='left', on=['anon_id', 'first_ed_time', 'admit_time'], suffixes=('_cohort', '_notes'))

# change first ED time to datetime
csn_notes.first_ed_time = pd.to_datetime(csn_notes.first_ed_time, utc=True)

csn_notes.head()

# make sure that the note's time is between first_ED_time and admit_time+24hours

filtered_csn_notes = csn_notes[(csn_notes.note_date_jittered_utc >= csn_notes.first_ed_time) &
                              (csn_notes.note_date_jittered_utc <= csn_notes.admit_time+timedelta(hours=24))]

# rename the csn that we're actually using
filtered_csn_notes = filtered_csn_notes.rename({'pat_enc_csn_id_coded_cohort':'pat_enc_csn_id_coded'}, axis=1)

filtered_csn_notes.pat_enc_csn_id_coded.nunique() # we have all of our csns

# Door to dispo

Get the notes that occur prior to admission

door_to_dispo_notes = filtered_csn_notes[filtered_csn_notes.note_date_jittered_utc < filtered_csn_notes.admit_time]

# some anon_ids do not have notes before admit (door_to_dispo window)
print(filtered_csn_notes.anon_id.nunique())
print(door_to_dispo_notes.anon_id.nunique())

# less than half of the notes occur before admit
print(filtered_csn_notes.shape) #(1567752, 17)
print(door_to_dispo_notes.shape) #(526164, 17)


door_to_dispo_notes.head(3)

# get the counts for all notes, rn notes, and md notes for this time window
def get_counts(window_notes, window):
    all_counts = window_notes.groupby(['pat_enc_csn_id_coded']).size().reset_index(name='{}_all_notes'.format(window))
    all_counts.head()

    # get the rn counts
    rn_counts = window_notes[window_notes.author_type == 'rn'].groupby(['pat_enc_csn_id_coded']).size().reset_index(name='{}_rn_notes'.format(window))
    rn_counts.head()

    # get the md counts
    md_counts = window_notes[window_notes.author_type == 'md'].groupby(['pat_enc_csn_id_coded']).size().reset_index(name='{}_md_notes'.format(window))
    md_counts.head()

    counts = all_counts.merge(rn_counts.merge(md_counts, how='outer'), how='outer')
    counts.head()
    
    return(counts)

# get door to dispo counts
door_to_dispo_counts = get_counts(door_to_dispo_notes, "door_to_dispo")
door_to_dispo_counts.head()

# Dispo to 24hr counts

Get the counts for dispo to 24hr: 

- effective_time >= admit_time

dispo_to_24hr_notes = filtered_csn_notes[filtered_csn_notes.note_date_jittered_utc >= filtered_csn_notes.admit_time]

# almost all anon_ids have notes during this window
print(filtered_csn_notes.anon_id.nunique())
print(dispo_to_24hr_notes.anon_id.nunique())

# majority of notes happen after admit
print(filtered_csn_notes.shape) #1567752
print(dispo_to_24hr_notes.shape) #1041588

dispo_to_24hr_counts = get_counts(dispo_to_24hr_notes, "dispo_to_24hr")
dispo_to_24hr_counts.head()

# join the two window counts together
all_counts = door_to_dispo_counts.merge(dispo_to_24hr_counts, how='outer')
all_counts.head()

# print(all_counts['door_to_dispo_rn_notes'].describe())
# print(all_counts['door_to_dispo_md_notes'].describe())
# print(all_counts['dispo_to_24hr_rn_notes'].describe())
# print(all_counts['dispo_to_24hr_md_notes'].describe())
all_counts.describe()

savedir

# save the output
savefile = "10_coh4_3hr_noOR_notes_counts.csv"
all_counts.to_csv(savedir + savefile, index=False)
