==> shc2021/11_Quality_controls_R.ipynb <==
# Quality control
- death is only accessed within the first 24 hours regardless of when the label time is

library(cowplot) #Installing package into 'C:/Users/User/Documents/R/win-library/4.0'
library(grid)
library(ggplot2)
library(ggExtra)
library(tidyverse)
library(lubridate)
library(gridExtra)
library(tableone)

# install.packages("tableone")

# directories
datadir = "../../DataTD/shc2021"
outdir = "../../OutputTD/shc2021"
figuredir = "../../OutputTD/Figures"
tabledir = "../../OutputTD/4_tables"

options(repr.matrix.max.rows=350, repr.matrix.max.cols=30)

coh4noOR <- read_csv(file.path(outdir, "7_cohort4_3hr_labels_noOR.csv"))
nOR = nrow(coh4noOR)
nOR
length(unique(coh4noOR$pat_enc_csn_id_coded))

coh4noORnoDKA <- read_csv(file.path(outdir, "7_cohort4_3hr_labels_noOR_noDKA.csv"))
N = nrow(coh4noORnoDKA)
N 
length(unique(coh4noORnoDKA$pat_enc_csn_id_coded))

names(coh4noORnoDKA)
head(coh4noOR, 2)

summary(coh4noOR)
summary(coh4noORnoDKA)

firstEDtime <- read_csv(file.path(outdir, "9_door_to_dispo_time.csv"))
nrow(firstEDtime)
length(unique(firstEDtime$pat_enc_csn_id_coded))

# 9_cohort4_3hr_noOR_first_ed_time is just a subset of 2 cols less without pat_csn and first_ed_time
firstEDtime2 <- read_csv(file.path(outdir, "9_cohort4_3hr_noOR_first_ed_time.csv"))
# nrow(firstEDtime2)
# length(unique(firstEDtime2$pat_enc_csn_id_coded))

head(firstEDtime, 3)

summary(firstEDtime)

# those with LOC changes within the first 24 hours since inpatient admission
# loc_change <- read_csv(file.path(outdir, "9_first_loc_change_time.csv")) # %>% mutate(admit_time = ymd_hms(admit_time))
all <- read_csv(file.path(outdir, "9_all_traj_time_team.csv")) # %>% mutate(admit_time = ymd_hms(admit_time))
nrow(all)
loc_change <- all %>% select(anon_id, pat_enc_csn_id_coded, hr_to_first_loc_change) %>% filter(!is.na(hr_to_first_loc_change)) #%>% distinct()
n_loc_change = nrow(loc_change) 
n_loc_change
nOR
n_loc_change/nOR * 100
length(unique(loc_change$pat_enc_csn_id_coded))

head(all, 1)
head(loc_change, 3)

loc_change <- coh4noOR %>% select(pat_enc_csn_id_coded, acute_to_critical_label_recent_3hr, critical_to_acute_label_recent_3hr,
                                  acute_to_critical_label_recent_24hr, critical_to_acute_label_recent_24hr) %>%
                    left_join(loc_change)
nrow(loc_change) 
length(unique(loc_change$pat_enc_csn_id_coded))

head(loc_change, 3)

loc_change %>% filter(acute_to_critical_label_recent_24hr == 1) %>% summary(first_loc_change_hours_since_admit)

loc_change %>% filter(acute_to_critical_label_recent_24hr == 0) %>% summary(first_loc_change_hours_since_admit)

notes <- read_csv(file.path(outdir, "10_coh4_3hr_noOR_notes_counts.csv"))
nrow(notes) #52532
summary(notes)

head(notes, 3)

### Remove labels from cohort and combine 2 new features: 
- Door-to-dispo time = hours_in_ed
- Number of notes in ed: rn, md, all

coh4noOR <- coh4noOR %>% select(anon_id, pat_enc_csn_id_coded, admit_time, first_label,
                                death_3hr_recent_label, death_6hr_recent_label, 
                                death_9hr_recent_label, death_12hr_recent_label, death_24hr_recent_label)
coh4noORnoDKA <- coh4noORnoDKA %>% select(anon_id, pat_enc_csn_id_coded, admit_time, first_label,
                                          death_3hr_recent_label, death_6hr_recent_label, death_9hr_recent_label, 
                                          death_12hr_recent_label, death_24hr_recent_label)

new_features <- coh4noOR %>% left_join(firstEDtime) %>% left_join(notes %>% select(-dispo_to_24hr_all_notes,
                                                                                   -dispo_to_24hr_rn_notes, 
                                                                                   -dispo_to_24hr_md_notes)) %>%
                select(-first_ed_time)
nrow(new_features)
names(new_features)
length(unique(new_features$pat_enc_csn_id_coded))

head(new_features, 3)

summary(new_features)

new_features <- new_features %>% mutate_if(is.numeric, ~replace(., is.na(.), 0))
summary(new_features)

# save the output
write.csv(new_features, file.path(outdir, "11_coh4_noOR_edtime_notecount.csv"), row.names=FALSE)

### COHORT4 no OR no DKA, and with main labels, 3 new features

cohort_new_features <- coh4noORnoDKA %>% left_join(new_features)
nrow(coh4noORnoDKA)
nrow(cohort_new_features)
length(cohort_new_features$pat_enc_csn_id_coded)
names(cohort_new_features)

# save the output
write.csv(cohort_new_features, file.path(outdir, "11_coh4_noOR_noDKA_edtime_notecount.csv"), row.names=FALSE)

### Other exploratory features:

all <- read_csv(file.path(outdir, "9_all_traj_time_team.csv"))
nrow(all) #52532

all[4013:4015,]

summary(all)
