==> shc2021/12_features_with_values_R.ipynb <==
### Descriptions:

JOIN all features with values together first, in the *long* format, to be merged with feature counts

**Inputs:** 
- `6_5_coh2_vitals`: has all processed vital signs (not the summary stats)
- `6_6_coh3_labs`: has all processed labs
- `6_8_coh4_all_imputedHWESI`: has demographics & imputed HW (cohort4_all) & imputed ESI (cohort3 and finalized w/ cohort4)
- `6_4_cohort4`: cohort3 with labels and not in old cohort

**Outputs:**
- `6_9_coh4_feature_values`: used cohort4, only new cohort, no overlapping!

### Importing R libraries

library(bigrquery)  # to query STARR-OMOP (stored in BigQuery) using SQL
library(tidyverse)
library(lubridate)
# library(mice)
# library(VIM) # for missing data plot

# library(data.table)
# library(Matrix)
# library(caret) # import this before glmnet to avoid rlang version problem
# library(glmnet)
# library(bit64)

# library(slam)
# library(mtools) for one hot coder, not available on Nero or use caret or tidyr
options(repr.matrix.max.rows=250, repr.matrix.max.cols=30)

### Call back all datasets: 
* demographic with imputed HW and ESI
* vitals, cohort with at least a complete set of vs only
* labs, cohort with at least a lab result
* cohort_final

# read inputs
datadir6 = "../../DataTD/shc2021"
outdir = "../../OutputTD/shc2021"

vitals0 <- read.csv(file.path(outdir, '5_coh2_vitals.csv')) 
nrow(vitals0) #1418197

labs0 <- read.csv(file.path(outdir, '6_coh3_labs.csv'))
nrow(labs0) #1887672

demos <- read.csv(file.path(outdir, '8_coh4_all_imputedHWESI.csv'))
nrow(demos) # all cohort4 52532

# cohort no OR
cohort4 <- read.csv(file.path(outdir, '11_coh4_noOR_edtime_notecount.csv'))
nrow(cohort4) # 51846

# # only new cohort NO DKA with new features (door-to-dispo time and note counts x3)
# cohort4 <- read.csv(file.path(outdir, '11_coh4_noOR_noDKA_edtime_notecount.csv'))
# nrow(cohort4) # 51846

names(cohort4)

# these come from different cohort
length(setdiff(vitals0$pat_enc_csn_id_coded, cohort4$pat_enc_csn_id_coded)) # 11094 different cohort
length(setdiff(cohort4$pat_enc_csn_id_coded, vitals0$pat_enc_csn_id_coded)) # 0

length(setdiff(vitals0$pat_enc_csn_id_coded, labs0$pat_enc_csn_id_coded)) # 5109 different cohort
length(setdiff(cohort4$pat_enc_csn_id_coded, labs0$pat_enc_csn_id_coded)) # 2593 less than before,, good

nrow(vitals0 %>% select(anon_id, pat_enc_csn_id_coded) %>% distinct())
length(unique(vitals0$pat_enc_csn_id_coded)) # 2nd cohort, larger 45613

nrow(demos %>% select(anon_id, pat_enc_csn_id_coded) %>% distinct())
length(unique(demos$pat_enc_csn_id_coded)) # 3nd cohort, 44258, using only those with a complete set of VS for ESI imputation

nrow(labs0 %>% select(anon_id, pat_enc_csn_id_coded) %>% distinct()) # 41627
length(unique(labs0$pat_enc_csn_id_coded)) # 41627 (similar before, labs have less than cohort)

nrow(cohort4 %>% select(anon_id, pat_enc_csn_id_coded) %>% distinct()) # 43980
length(unique(cohort4$pat_enc_csn_id_coded)) # 41627 (similar before, labs have less than cohort)

length(setdiff(demos$pat_enc_csn_id_coded, cohort4$pat_enc_csn_id_coded)) # 278
length(setdiff(vitals0$pat_enc_csn_id_coded, cohort4$pat_enc_csn_id_coded)) # 1633
length(setdiff(labs0$pat_enc_csn_id_coded, cohort4$pat_enc_csn_id_coded)) # 261

head(cohort4, 1)
head(demos, 1)
head(vitals0, 1)
head(labs0, 1)

colnames(cohort4)
colnames(demos)
colnames(vitals0)
colnames(labs0)

# cohort <- cohort %>% mutate(admit_time = ymd_hms(admit_time)) 

# cohort <- cohort4 %>% select(anon_id, pat_enc_csn_id_coded, admit_time, first_label, death_24hr_recent_label, death_24hr_max_label)
cohort <- cohort4 %>% select(anon_id, pat_enc_csn_id_coded, admit_time, first_label,
                             death_3hr_recent_label, death_6hr_recent_label, death_9hr_recent_label,
                             death_12hr_recent_label, death_24hr_recent_label)
nrow(cohort)

time_count <- cohort4 %>% select(anon_id, pat_enc_csn_id_coded, ed_hour = hours_in_ed, 
                                 rn_note=door_to_dispo_rn_notes, md_note=door_to_dispo_md_notes,
                                 all_note=door_to_dispo_all_notes) %>%
                mutate(feature_type = 'edtime_notecount')
nrow(time_count)
sum(is.na(time_count)) # na was converted to 0, no notes
summary(time_count)

# turn to long format
tc_long <- gather(time_count, features, values, ed_hour:all_note, factor_key=TRUE) %>%
                mutate(feature_type = "time_count")
nrow(tc_long)
nrow(time_count) * 4

# turn the wide format to long for demographics
demo_long <- gather(demos, features, values, ESI_i:race.White, factor_key=TRUE) %>%
                mutate(feature_type = "demo") %>% 
                select(-admit_time, -label, -inpatient_data_id_coded) %>% #mutate(admit_time = ymd_hms(admit_time)) %>% 
             right_join(cohort)

sum(is.na(demos))
nrow(demo_long) # = nrow(cohort) *17 = 881382
nrow(cohort) * 17

head(demo_long, 2)

# drop admit_time from vitals and labs because they might be in different formats utc -- mess up the merge
vitals <- vitals0 %>% select(anon_id, pat_enc_csn_id_coded, features, values, feature_type, 
                             time = recorded_time) %>% #mutate(admit_time = ymd_hms(admit_time)) %>%
                         right_join(cohort) #%>% drop_na()
labs <- labs0 %>% select(anon_id, pat_enc_csn_id_coded, features, values, feature_type, 
                         time = result_time) %>% #mutate(admit_time = ymd_hms(admit_time)) %>% 
                    right_join(cohort) %>% drop_na() # due to csn without labs in cohort4

# JCdcafca and 131187786922, 20015518, 0, 2016-05-15 20:57:00+00:00 
# in the no_labels cohort (and Conor's cohort) but not in the updated cohort
noinco <- cohort[cohort$pat_enc_csn_id_coded %in% no_labels$pat_enc_csn_id_coded,]
head(no_labels[!no_labels$pat_enc_csn_id_coded %in% noinco$pat_enc_csn_id_coded, ])

length(unique(demo_long$pat_enc_csn_id_coded))
length(unique(vitals$pat_enc_csn_id_coded))
length(unique(labs$pat_enc_csn_id_coded)) # 49253 due to drop NA
length(unique(tc_long$pat_enc_csn_id_coded)) # 51846

head(demo_long, n=1)
head(vitals, n=1)
head(labs, n=1)
head(tc_long, 1)

summary(tc_long$values)
summary(demo_long$values)
summary(vitals$values)
summary(labs$values) # negatives

# combine demos, vitals and labs, long format, with "time"
feats <- bind_rows(demo_long, tc_long, vitals, labs)
feats <- as.data.frame(unclass(feats))

nrow(feats) #3711563
nrow(feats %>% select(anon_id, pat_enc_csn_id_coded) %>% distinct()) # 51846
nrow(feats %>% select(pat_enc_csn_id_coded) %>% distinct()) # 51846

feats %>% count(feature_type) # shouldn't have NA
summary(feats$values) 

feats %>% group_by(feature_type, features) %>% count()

names(feats)

# rearrange columns
cohort_feats <- feats %>% select(anon_id, pat_enc_csn_id_coded, admit_time, first_label, 
                                 death_3hr_recent_label, death_6hr_recent_label, death_9hr_recent_label, 
                                 death_12hr_recent_label, death_24hr_recent_label,
                                 feature_type, features, values, time)
nrow(cohort_feats) # 3867101 
length(unique(cohort_feats$pat_enc_csn_id_coded)) # 51846 vs 52532

head(cohort_feats, 1)
tail(cohort_feats, 1)

# only have new cohort
# write.csv(cohort_feats, file.path(outdir, "12_coh4_noOR_noDKA_feature_values.csv"), row.names=FALSE)
write.csv(cohort_feats, file.path(outdir, "12_coh4_noOR_feature_values.csv"), row.names=FALSE)

featuredir = "../../OutputTD/2_features"
cohort4_0_feats <- read.csv(file.path(featuredir, "2_7_coh4_feature_values.csv"))
nrow(cohort4_0_feats) # 3085046
length(unique(cohort4_0_feats$pat_enc_csn_id_coded))

# NO overlapping! sanity check
length(setdiff(cohort4_feats$pat_enc_csn_id_coded, cohort4_0$pat_enc_csn_id_coded))
length(setdiff(cohort4_0_feats$pat_enc_csn_id_coded, cohort4_feats$pat_enc_csn_id_coded))
