==> shc2021/13.2_featurize_vitals_labs_values.ipynb <==
This notebook was developed by Gautam Machiraju and Conor Corbin, modified by Minh Nguyen

### Description: major changes
Because binning relies on training data, we need to *use all the old and new data* here
- Use 2015 - *2018* as the first training data to check for value distributions with 10 quantiles 
- Use these distributions to assign new validation and test values to quantile bins 
- 2 major steps:
    - With the old data, use 2015 - 2018 as the first training data to create quantile distribution, and 2019 - 03/2020 as validation data. This is for modeling training to tune hyperparameters
    - Comebine the new data with old data, use 2015 - 03/2020 as train/val data to create quantile distribution, and 04/2020 - 2021 as the test data, assign their binns from the the train/val data.
    
**Input:** 
- `6_9_coh4_feature_values`: used new cohort4
- `2_7_coh4_feature_values`: used old cohort4

**Outputs:** `...coh4...`
- `6_10_binned_labs_vitals_train`:
    - use 2015 - *2018* as trained data for value distributions to bin validation data of 2019-03/2020
    - this is for the purpose of train data and select model hyper parameters on validation data
    - validation data > 2018 (2019 - 03/2020) 
- `6_10_binned_labs_vitals_test`:
    - after tuning hyperparameters, the value distributions are done again on data trai/val data from 2015 - 03/2020
    - test data (04/2020 - 2021) are binned on based on this distribution
    - binned test data is used for the final prediction and evaluation of the new model after THICK DESCRIPTION work
- Note that the size of these 2 data sets are different

import os 
import numpy as np
import pandas as pd
from datetime import datetime, timedelta

import os 
from google.cloud import bigquery
from google.cloud.bigquery import dbapi

##Use correct path based on whether you are, Nero or local
# use Ctrl + Insert to copy and Shift + Insert to paste

# for Nero:
# os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = '/home/minh084/.config/gcloud/application_default_credentials.json' 
os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = '/home/jupyter/.config/gcloud/application_default_credentials.json'

# for local computer:
# os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = r'C:\Users\User\AppData\Roaming\gcloud\application_default_credentials.json' 

##set correct Nero project
os.environ['GCLOUD_PROJECT'] = 'som-nero-phi-jonc101' 

datadir = "../../DataTD/shc2021"
outdir = "../../OutputTD/shc2021"

pd.set_option('display.max_columns', 50)
pd.set_option('display.max_rows', 50)

### Train And Apply Featurizer - change for new validation data after March 2020! 

#### Depending on which dataset used for train/val or train_val/test, choose different time split
- For the purpose of training to select hyperparameters for models, use 2015 - 2017 (<2018) for binning distribution --> **NEW: use 2015 - 2018**
- One a final model with hyperparameters are selected, for the final prediction on an unseen test set, use 2015 - 2018 for binning distribution ( < 2019) --> **NEW: use 2015 - 03/2020**
- basically with this new validation data, we could use all the old data in this process, without saving the 2019 - 03/2020 as the final test set

# either pull from BQ
query = """select * from triageTD.6_9_coh4_feature_values.csv"""
query_job = client.query(query)
df = query_job.result().to_dataframe()

# df = pd.read_csv(os.path.join(outdir, "12_coh4_noOR_noDKA_feature_values.csv"))
df = pd.read_csv(os.path.join(outdir, "12_coh4_noOR_feature_values.csv"))

# need this if read from local file, if pulled from BQ, already in datetime format
df["admit_time"] = pd.to_datetime(df["admit_time"], utc=True) 

print(len(df)) # 3,085,046 

# check to see if any feature is null
df[df['features'].isnull()]

# df.loc[df['pat_enc_csn_id_coded'] == 131250899044]

df = df[df['feature_type'].isin(['labs', 'vitals'])]
df["time"] = pd.to_datetime(df["time"])

print(len(df)) #2337386

df.dtypes

# calculate time difference in hours between admit time and time of vitals and labs,
df['difftime'] = round((df['admit_time'] - df['time']).dt.total_seconds()/3600, 1)

# check for observations more than 24 hours before admission time
print(len(df.loc[df['difftime'] > 24])) # 40070
df[['difftime']].describe()

# remove those observations:
df = df.loc[df['difftime'] <= 24]
print(len(df)) #2297316
print(df.pat_enc_csn_id_coded.nunique()) # lost 1

# check for coh4:
2337386 - 40070

### Process Continuous Features
Create a function that "trains" binning featurizer (computes distribution of values) based on subset of the data.  This is important because we only want to build the distribution with our training set and apply the bin featurizer
to the test set (prevents leakage). 

Then create a function that "applies" the trained featurizer on a set of data. 


def convert_to_dict(look_up_table):
    """Converts df look up table to dictionary for faster look up later"""
    bin_val_dict = {}
    for feature in look_up_table['features'].unique():
        feature_bin_vals = look_up_table[look_up_table['features'] == feature]
        for _bin in feature_bin_vals['bins'].unique():
            if feature not in bin_val_dict:
                bin_val_dict[feature] = {}
                bin_val_dict[feature]['min'] = []
                bin_val_dict[feature]['max'] = []

            min_val_for_bin = feature_bin_vals[feature_bin_vals['bins'] == _bin]['values']['min'].values[0]
            max_val_for_bin = feature_bin_vals[feature_bin_vals['bins'] == _bin]['values']['max'].values[0]

            bin_val_dict[feature]['min'].append(min_val_for_bin)
            bin_val_dict[feature]['max'].append(max_val_for_bin)
    return bin_val_dict

    
def train_featurizer(df_train):
    """
    Compute percent_ranks and generates a look up table of min and max bin values
    Input : long form dataframe with features and values where values are the continuous values of labs / vitals
    Output: look up table - dict of dict of lists (key1 = feature_name, key2 = max or min, values = lists of values)
    """
    # Compute percentiles and bins
    df_train['percentiles'] = df_train.groupby('features')['values'].transform(lambda x: x.rank(pct=True))
    df_train['bins'] = df_train['percentiles'].apply(lambda x: int(x * 10))
    
    # Generate look up table and conver to dictionary stucture
    look_up_table_df = df_train.groupby(['features', 'bins']).agg({'values' : ['min', 'max']}).reset_index()
    look_up_table = convert_to_dict(look_up_table_df)
    
    ### Sanity Check. Ensure that min vector for each feature is strictly increasing (no ties!)
    # Should be the case because ties are given same percentile rank in default pandas rank function
    for feature in look_up_table:
        mins = look_up_table[feature]['min']
        for i in range(len(mins)-1):
            assert mins[i] < mins[i+1]
    
    return look_up_table


def apply_featurizer(df, look_up_table):
    
    def get_appropriate_bin(feature, value, look_up_table):
        """Takes in feature, value and look up table and returns appropriate bin

        Quick Note: For some features, we do not have 10 bins.  This happens when we have many many ties in the 
        percent rank - and the percent rank alg returns ties as the average rank within that tie. So for instance
        we're trying to break each feature up into deciles where each bin covers range of 10% of the examples. But if more
        than 10% of the examples take on 1 value - then bins can be skipped. This shouldn't really be a problem
        for downstream tasks - just something to be aware of. This also means 'bins' and 'bins_applied' won't have
        perfect overlap in features that end up having less than 10 bins

        """
        mins = look_up_table[feature]['min']
        for i in range(len(mins) - 1):
            # If value is smaller than min value of smallest bin (in test time) - then return 0 (smallest bin)
            if i == 0 and value < mins[i]:
                return i

            if value >= mins[i] and value < mins[i+1] :
                return i

        # Then in last bin
        return len(mins)-1
    
    df['bins_applied'] = df[['features', 'values']].apply(
        lambda x: get_appropriate_bin(x['features'], x['values'], look_up_table), axis=1)
    
    return df

### Train And Apply Featurizer - change for new validation data after March 2020! 

#### Depending on which dataset used for train/val or train_val/test, choose different time split
- For the purpose of training to select hyperparameters for models, use 2015 - 2017 (<2018) for binning distribution --> **NEW: use 2015 - 2018**
- One a final model with hyperparameters are selected, for the final prediction on an unseen test set, use 2015 - 2018 for binning distribution ( < 2019) --> **NEW: use 2015 - 03/2020**
- basically with this new validation data, we could use all the old data in this process, without saving the 2019 - 03/2020 as the final test set

# for train/val VS. train_val/test:
# df_train =  df[(df['admit_time'].dt.year < 2019)] # old 2018
# df_train =  df[(df['admit_time'].dt.year < 2021)] # old 2019

# FOR TRAINING/VALIDATION purpose - chosing hyperparameters, 
# do decile distribution only on train set
df_train =  df[(df['admit_time'].dt.year < 2019)]

### OR TRAIN_VAL/TEST
# AFTER training and chosing hyperparameters, RE-DO DECILE distribution on both train_val set (2015 - 2018)
df_train =  df[(df['admit_time'].dt.year < 2021)] # all up to 03/2020, this used the old dataset

# continue here
look_up_table = train_featurizer(df_train)
df_featurized = apply_featurizer(df, look_up_table)

# sanity check of the total cohort size with time splitting
# Change the years
print(len(df))
print(len(df[(df['admit_time'].dt.year < 2019)]))
print(len(df[(df['admit_time'].dt.year == 2019)]))
print(len(df[(df['admit_time'].dt.year > 2019)]))
len(df[(df['admit_time'].dt.year < 2019)]) + len(df[(df['admit_time'].dt.year == 2019)]) + len(df[(df['admit_time'].dt.year > 2019)])

### Quick Sanity Check
For features that have 10 bins from 0 to 9 - `bins` should be same as `bins_applied`

df_train = apply_featurizer(df_train, look_up_table)
look_up_table_df = df_train.groupby(['features', 'bins']).agg({'values' : ['min', 'max']}).reset_index()

features_with_0_9_bins = []
for feature in look_up_table_df:
    num_bins = len(look_up_table_df[look_up_table_df['features'] == feature]['bins'].values)
    ten_in_bins = 10 in look_up_table_df[look_up_table_df['features'] == feature]['bins'].values
    if num_bins == 10 and not ten_in_bins:
        features_with_0_9_bins.append(feature)

for feature in features_with_0_9_bins:
    df_test = df_train[df_train['features'] == 'feature']
    for b_real, b_computed in zip(df_test['bins'].values, df_test['bins_applied'].values):
        assert(b_real == b_computed)

# checked against R 2.9_
len(df_train) # 1678328 --old coh4: 1254522 vs 1714603 

### Little bit of house cleaning
Create new feature names that reflect which bin the value belongs in

columns = ['anon_id', 'pat_enc_csn_id_coded', 'admit_time', 'feature_type', 'features', 'values', 'bins_applied']
df_new = df_featurized[columns]

df_new['features'] = ['_'.join([x, str(y)]) for x, y in zip(df_new['features'].values, df_new['bins_applied'].values)] 

len(df_new) #coh4 2337386 

### Get Counts representation
Group by patient, cns, and feature name (with bin value appended to feature name) and make value the number of times
that particular feature appears for that csn id. 

df_final = df_new.groupby(['anon_id', 'pat_enc_csn_id_coded', 'features']).agg(
    {'admit_time' : 'first',
     'feature_type' : 'first',
     'values' : 'count'}).reset_index()

columns = ['anon_id', 'pat_enc_csn_id_coded', 'admit_time', 'feature_type', 'features', 'values']
df_final = df_final[columns] # reorder columns
 
df_final['feature_type'] = [x + '_results' if x == 'labs' else x for x in df_final['feature_type'].values]

#### Depending on which dataset used for train/val or train_val/test, name: _train vs _test
- Rename feature_type to reflect training set used: ..._train or ...test
- 'vitals_train' means everything up to 2017 (**NEW- 2018**) used. (train vs val)
- 'vitals_test' means everything up to 2018 (**NEW - 02/2020**) used. (train + val vs test)

# for train/val to select hyperparameters, binned distributions on train/val: 
# df_train =  df[(df['admit_time'].dt.year < 2018)]
df_final['feature_type'] = [x + '_train' for x in df_final['feature_type'].values]

### OR 2nd time on train_val/test: 
# df_train =  df[(df['admit_time'].dt.year < 2019)]
df_final['feature_type'] = [x + '_test' for x in df_final['feature_type'].values]

print(len(df_final)) # old - updated coh4 1825210(train) vs 1826919 (train_val)
print(len(df)) # updated coh4 2297316 same

# Sanity check - sum of the counts should be length of the orginal dataframe
assert df_final['values'].sum() == len(df)

df_final.head(10)

### SAVE, depending on which cohort and which dataset

#### cohort4, train/val, train_val/test

# for training to select hyperparameters: df_train =  df[(df['admit_time'].dt.year < 2019)] - new
# df_final.to_csv(os.path.join(outdir, '13.2_coh4_binned_labs_vitals_train.csv'), index=False)
df_final.to_csv(os.path.join(outdir, '13.2_coh4_noOR_binned_labs_vitals_train.csv'), index=False)

### OR
# after selecting hyperparameters, re-do binned distributions: df_train =  df[(df['admit_time'].dt.year < 2019)]
# df_final.to_csv(os.path.join(featuredir, '2_8_coh4_binned_labs_vitals_test.csv'), index=False)

### ReDO with TRAIN_VAL/TEST

- First, format the new data 04/2020 - 2021
- Remove the bins_applied column from the old data
- Concat the 2 data into one for re-binning

### OR 2nd time on TRAIN_VAL/TEST
# AFTER training and chosing hyperparameters, RE-DO DECILE distribution on both train_val set (2015 - 2018)
df_train =  df[(df['admit_time'].dt.year < 2020) | 
            (df['admit_time'].dt.year == 2020) & (df['admit_time'].dt.month < 4)] # all old data up to 03/2020

# df_train = df0 # all old data 2015 - 03/2020, not including any new data for untouched test set, look up afterwards
print(len(df_train)) # 2297316

# continue here
look_up_table = train_featurizer(df_train)
df_featurized = apply_featurizer(df, look_up_table)  # NOTE, df here is the combined df0 and df21

# sanity check
df_train = apply_featurizer(df_train, look_up_table)
look_up_table_df = df_train.groupby(['features', 'bins']).agg({'values' : ['min', 'max']}).reset_index()

features_with_0_9_bins = []
for feature in look_up_table_df:
    num_bins = len(look_up_table_df[look_up_table_df['features'] == feature]['bins'].values)
    ten_in_bins = 10 in look_up_table_df[look_up_table_df['features'] == feature]['bins'].values
    if num_bins == 10 and not ten_in_bins:
        features_with_0_9_bins.append(feature)

for feature in features_with_0_9_bins:
    df_test = df_train[df_train['features'] == 'feature']
    for b_real, b_computed in zip(df_test['bins'].values, df_test['bins_applied'].values):
        assert(b_real == b_computed)

# cleaning - 
# create new feature names that reflect which bin the value belongs in
columns = ['anon_id', 'pat_enc_csn_id_coded', 'admit_time', 'feature_type', 'features', 'values', 'bins_applied']
df_new = df_featurized[columns]

df_new['features'] = ['_'.join([x, str(y)]) for x, y in zip(df_new['features'].values, df_new['bins_applied'].values)] 

print(len(df_new)) #3186016 vs 2780923 noOR

# get count presentation
df_final = df_new.groupby(['anon_id', 'pat_enc_csn_id_coded', 'features']).agg(
    {'admit_time' : 'first',
     'feature_type' : 'first',
     'values' : 'count'}).reset_index()

columns = ['anon_id', 'pat_enc_csn_id_coded', 'admit_time', 'feature_type', 'features', 'values']
df_final = df_final[columns] # reorder columns

df_final['feature_type'] = [x + '_results' if x == 'labs' else x for x in df_final['feature_type'].values]

# OR 2nd time on train-val/test: rename feature type to reflect train or test
df_final['feature_type'] = [x + '_test' for x in df_final['feature_type'].values]

print(len(df_final)) # 2534674 vs 2220732 noOR
print(len(df)) # 3186016 vs 2780923 noOR

# sanity check
assert df_final['values'].sum() == len(df)

### OR 2nd time on train-val/test
# df_final.to_csv(os.path.join(outdir, '13.2_coh4_binned_labs_vitals_test.csv'), index=False)
df_final.to_csv(os.path.join(outdir, '13.2_coh4_noOR_binned_labs_vitals_test.csv'), index=False)

df_final.pat_enc_csn_id_coded.nunique()

### View

for feature in look_up_table_df['features'].unique():
    print(look_up_table_df[look_up_table_df['features'] == feature])
