==> shc2021/14_featurizedvalues_ordercounts_R.ipynb <==
### Description:
**REMOVE 1 patient visit = 131289492397 from the final cohort4**
- Read in the decile-binned values of labs and vitals
- Read in the data `coh4_order_code_counts` from sql order counts, combine Microbiology Culture with Microbiology 
- Combine this featurized data for full complex data modeling
- Create a smaller random sample for testing 
    
**Inputs**: 
- `6_10_coh4_binned_labs_vitals_train`: from feature_values, binned into deciles, 2015 - 03/2020
- `6_10_coh4_binned_labs_vitals_test`: 2015 - 03/2020 + 04/2020 - 2021
- `coh4_order_code_counts` and `coh4_order_code_counts_2021`: (from SQL) queried counts for different orders
    
**Outputs**: 
- have demo, vitals and labs in binned counts, order counts: long format, with year extracted
- `6_11_coh4_features_all_long_year`: will be input to 3_models, sparse_matrix.py
    - 2015 - 2018 as trained data for value distributions to bin validation data of 2018
        - used for training and select model hyperparameters based on validation data
    - test data 2019 - 03/2020 left unused
    - after training and selecting hyperparameters, re-do value distribution on 2015 - 03/2020 (train + val) data
    - test data (04/2020 - 2021) are now binned on based on this new train/val distribution
    - binned test data is used for the final prediction and evaluation of new model after THICK DESCRIPTION work


library(data.table)
library(tidyverse)
library(lubridate)
# library(Matrix)
# library(slam)
# library(bit64)
# library(mtools) for one hot coder, not available on Nero or use caret or tidyr
options(repr.matrix.max.rows=200, repr.matrix.max.cols=30)

# install.packages(c("dbplyr", "RSQLite"))

datadir = "../../DataTD/shc2021"
outdir = "../../OutputTD/shc2021"

# cohort4 <- read.csv(file.path(outdir, '11_coh4_noOR_noDKA_edtime_notecount.csv'))
cohort4 <- read.csv(file.path(outdir, '11_coh4_noOR_edtime_notecount.csv'))

nrow(cohort4)

cohort <- cohort4 %>% select(anon_id, pat_enc_csn_id_coded, admit_time, first_label,
                             death_3hr_recent_label, death_6hr_recent_label, death_9hr_recent_label,
                             death_12hr_recent_label, death_24hr_recent_label)
nrow(cohort)

# Cold_Counts_val.sql under ../SQL/feature_counts -- OLD

# Sequence of codes assigned to patients 

WITH sequence as (
-- # Get (lab, procedure, imaging and microbiology orders from order_proc) within one year prior to current admission time
SELECT DISTINCT lab.order_proc_id_coded as order_id,
                lab.order_type as feature_type,
                lab.description as features,
                cohort.anon_id, cohort.pat_enc_csn_id_coded, cohort.admit_time, 
                cohort.first_label, cohort.death_24hr_recent_label
    
FROM triageTD.6_7_cohort4 as cohort
INNER JOIN shc_core_2021.order_proc as lab
USING (anon_id)
WHERE 
(lab.order_type in ('Imaging', 'Procedures', 'Lab')
AND lab.order_time_jittered_utc < cohort.admit_time
AND timestamp_add(lab.order_time_jittered_utc, INTERVAL 24*365 HOUR) >= cohort.admit_time)
OR
(lab.order_type in ('Microbiology', 'Microbiology Culture')
AND lab.order_time_jittered_utc < cohort.admit_time
AND lab.pat_enc_csn_id_coded = cohort.pat_enc_csn_id_coded)
UNION DISTINCT

-- # Get icd10 codes from full timeline, from all previous admissions
SELECT DISTINCT pat_enc_csn_id_coded as order_id, -- fill in because dx table doesn't have order_id
                'Diagnosis' as feature_type, dx.icd10 as features,
                cohort.anon_id, cohort.pat_enc_csn_id_coded, cohort.admit_time, 
                cohort.first_label, cohort.death_24hr_recent_label
    
FROM triageTD.6_7_cohort4 as cohort
INNER JOIN shc_core_2021.diagnosis as dx
USING (anon_id)
WHERE cohort.pat_enc_csn_id_coded <> dx.pat_enc_csn_id_jittered
AND dx.start_date_utc < cohort.admit_time 
AND dx.icd10 is not NULL
UNION DISTINCT

-- # Get med orders from order_med from within one year prior to current admission time
SELECT DISTINCT med.order_med_id_coded order_id,
                'Meds' as feature_type, med.med_description as features,
                cohort.anon_id, cohort.pat_enc_csn_id_coded, cohort.admit_time, 
                cohort.first_label, cohort.death_24hr_recent_label
    
FROM triageTD.6_7_cohort4 as cohort
INNER JOIN shc_core_2021.order_med as med
USING (anon_id)
WHERE med.order_start_time_utc < cohort.admit_time
AND timestamp_add(med.order_start_time_utc, INTERVAL 24*365 HOUR) >= cohort.admit_time
AND med.order_start_time_utc IS NOT NULL) -- maybe redundent but fine

-- # Group by feature and encounter, and count # distinct orders/assignemnts of the feature
SELECT anon_id, pat_enc_csn_id_coded, admit_time, feature_type, features, COUNT (DISTINCT order_id) as values
FROM sequence
GROUP BY anon_id, pat_enc_csn_id_coded, admit_time, feature_type, features
ORDER BY pat_enc_csn_id_coded

# this one used cohort noOR only to query orders
orders <- read.csv(file.path(datadir, 'coh4_noOR_order_code_counts_2021.csv'))
nrow(orders) #8333201

head(orders, 1)

orders %>% group_by(feature_type) %>% count(sort=TRUE)

orders_micro <- orders %>% mutate(feature_type = ifelse(str_detect(feature_type, "Microbiology"), "Microbiology", feature_type))
orders_micro %>% group_by(feature_type) %>% count(sort=TRUE)

orders_micro %>% group_by(feature_type) %>% summarise(avg = mean(values), sd = sd(values))

length(unique(orders_micro$pat_enc_csn_id_coded)) # 52525
length(setdiff(orders_micro$pat_enc_csn_id_coded, cohort$pat_enc_csn_id_coded)) # 686 # ok because this is from cohort noOR
length(setdiff(cohort$pat_enc_csn_id_coded, orders_micro$pat_enc_csn_id_coded)) # 7
# not everyone has orders
# for the new cohort, every has orders

head(orders_micro, 1)
head(cohort, 1)

orders_micro <- orders_micro %>% filter(pat_enc_csn_id_coded %in% cohort$pat_enc_csn_id_coded)
length(unique(orders_micro$pat_enc_csn_id_coded)) # 51839 (7 missing) vs

### Get the demographic features from each cohort, then combine

#### Run the old cohort

# check all features with values
# values <- read.csv(file.path(outdir, '12_coh4_noOR_noDKA_feature_values.csv'))
values <- read.csv(file.path(outdir, '12_coh4_noOR_feature_values.csv'))

nrow(values) # coh4 3867101 
length(unique(values$pat_enc_csn_id_coded)) # 51846 
values %>% group_by(feature_type) %>% count()

# get demographics features
demos <- values %>% filter(feature_type == 'demo') %>% select(anon_id, pat_enc_csn_id_coded, admit_time, 
                                                              feature_type, features, values)
nrow(demos) # coh4 881382 
length(unique(demos$pat_enc_csn_id_coded)) # 51846

demos %>% group_by(feature_type) %>% count()
demos %>% group_by(features) %>% count()

### Check the featurized binned labs_vitals
- Keep them both (`..._train` and `..._test`) in the same dataframe
- Run `2_7_feature_values`, `2_8_binned_labs_vitals_train`, and `2_8_binned_labs_vitals_test` 
- Rerun the old cohort and new cohort separately, then combine the demos and order counts features
- The featurirzed labs and vitals are already combined in the previous notebook 6.10

# read in train_binned labs and vitals
# trainbinned_labs_vitals <- read.csv(file.path(outdir, '13_coh4_binned_labs_vitals_train.csv'))
# trainbinned_labs_vitals <- read.csv(file.path(outdir, '13.2_coh4_binned_labs_vitals_train.csv'))

### different cohort no OR only
# trainbinned_labs_vitals <- read.csv(file.path(outdir, '13_coh4_noOR_binned_labs_vitals_train.csv'))
trainbinned_labs_vitals <- read.csv(file.path(outdir, '13.2_coh4_noOR_binned_labs_vitals_train.csv'))


nrow(trainbinned_labs_vitals) #coh4: old 2185308
length(unique(trainbinned_labs_vitals$pat_enc_csn_id_coded)) # 51845
trainbinned_labs_vitals %>% group_by(feature_type) %>% count()

# this patient has no vitals or labs within 24 hours? difftime in featurized didn't take this one!!!! removed 
setdiff(cohort$pat_enc_csn_id_coded, trainbinned_labs_vitals$pat_enc_csn_id_coded)

# read in test_binned labs and vitals, new cohort
# testbinned_labs_vitals <- read.csv(file.path(outdir, '13_coh4_binned_labs_vitals_test.csv'))
# testbinned_labs_vitals <- read.csv(file.path(outdir, '13.2_coh4_binned_labs_vitals_test.csv'))

# COHORT noOR only
# testbinned_labs_vitals <- read.csv(file.path(outdir, '13_coh4_noOR_binned_labs_vitals_test.csv'))
testbinned_labs_vitals <- read.csv(file.path(outdir, '13.2_coh4_noOR_binned_labs_vitals_test.csv'))

nrow(testbinned_labs_vitals) #coh4 2184675 
length(unique(testbinned_labs_vitals$pat_enc_csn_id_coded)) # 51845
testbinned_labs_vitals %>% group_by(feature_type) %>% count()

# read in test_binned labs and vitals, new cohort
testbinned_labs_vitals1 <- read.csv(file.path(outdir, '13_coh4_binned_labs_vitals_test.csv'))

nrow(testbinned_labs_vitals1) #coh4 2184675 

# read in test_binned labs and vitals, new cohort
testbinned_labs_vitals2 <- read.csv(file.path(outdir, '13_coh4_all_binned_labs_vitals_test.csv'))

nrow(testbinned_labs_vitals2) #coh4 2184675 
all.equal(testbinned_labs_vitals2, testbinned_labs_vitals1)

head(trainbinned_labs_vitals, 1)
head(testbinned_labs_vitals, 1)

### Combine all features

head(demos, 1)
head(orders_micro,1)
head(trainbinned_labs_vitals, 1)
head(testbinned_labs_vitals, 1)

# combine all features for the complex data set, remove time column since demo doesn't have it
features_all <- bind_rows(demos, orders_micro, trainbinned_labs_vitals, testbinned_labs_vitals) %>%
                    filter(pat_enc_csn_id_coded != 131289492397) %>% mutate(admit_time = ymd_hms(admit_time))
nrow(features_all) # 14269242 
length(unique(features_all$pat_enc_csn_id_coded)) # 51845
length(unique(features_all$features)) # 49278
features_all %>% group_by(feature_type) %>% count(sort=TRUE)

head(features_all, 1)
tail(features_all, 1)

features_all %>% filter(pat_enc_csn_id_coded == 131062572931) %>% distinct(admit_time)

# sanity check 
n2 <- features_all %>% select(feature_type, features) %>% distinct() %>% 
        group_by(features) %>% count(sort=TRUE) %>% filter(n==2)
head(n2)
tail(n2)

features_all %>% filter(features %in% n2$features) %>% distinct(feature_type)# group_by(features, feature_type) %>% count()

features_all %>% group_by(feature_type, features) %>% count(sort=TRUE) %>% head(30)

names(features_all)
nrow(features_all) # 13584533
summary(features_all$values)
sum(is.na(features_all$values))

features_all_year <- features_all %>% mutate(year = year(admit_time)) #, month = month(admit_time))
features_all_year %>% group_by(year) %>% count()
# features_all_year %>% group_by(month) %>% count()

cohort %>% mutate(year = year(admit_time)) %>% group_by(year) %>% count()

nrow(features_all_year)
length(unique(features_all_year$pat_enc_csn_id_coded)) # 60464
length(setdiff(features_all_year$pat_enc_csn_id_coded, cohort$pat_enc_csn_id_coded))
length(setdiff(cohort$pat_enc_csn_id_coded, features_all_year$pat_enc_csn_id_coded)) # has not remove the 1 ptvisit yet

head(cohort,3)

head(features_all_year, 1)

cohort_final <- cohort %>% filter(pat_enc_csn_id_coded != 131289492397)
nrow(cohort_final)
summary(cohort_final)

length(setdiff(features_all_year$pat_enc_csn_id_coded, cohort_final$pat_enc_csn_id_coded)) # 686 # ok because this is from cohort noOR
length(setdiff(cohort_final$pat_enc_csn_id_coded, features_all_year$pat_enc_csn_id_coded)) # 7

# FINAL COHORT no OR no DKA, remove 1

# remove 1 single csn, noOR, noDKA
# write.csv(cohort_final, file = file.path(outdir, "14_cohort_final.csv"), row.names=FALSE) 

# only noOR
write.csv(cohort_final, file = file.path(outdir, "14_cohort_noOR_final.csv"), row.names=FALSE) 

# write.csv(features_all_year, file = file.path(outdir, "14_coh_all_features_all_long_year.csv"), row.names=FALSE)
# write.csv(features_all_year, file = file.path(outdir, "14_2_coh_all_features_all_long_year.csv"), row.names=FALSE) 


# different cohort, noOR only
# write.csv(features_all_year, file = file.path(outdir, "14_coh_noOR_all_features_all_long_year.csv"), row.names=FALSE)
write.csv(features_all_year, file = file.path(outdir, "14_2_coh_noOR_all_features_all_long_year.csv"), row.names=FALSE) 

summary(features_all_year)
